Added artists field to User PATCH

+++ proyecto_musica/authentication/serializers.py
@@ -14,10 +14,6 @@ class RegisterSerializer(serializers.ModelSerializer):
         model=User
         fields=('username','email','password','skills','genres','artists')
 
-    def get_artists(self, obj):
-        artists = self.context.get("artists")
-        return get_list_field(None, obj.email, "artist", artists)
-
     def get_skills(self, obj):
         skills = self.context.get("skills")        
         return get_list_field(None, obj.email, "skill", skills)
@@ -26,6 +22,10 @@ class RegisterSerializer(serializers.ModelSerializer):
         genres = self.context.get("genres")        
         return get_list_field(None, obj.email, "genre", genres)
 
+    def get_artists(self, obj):
+        artists = self.context.get("artists")
+        return get_list_field(None, obj.email, "artist", artists)
+    
     def create(self, validated_data):
         user =  User.objects.create_user(**validated_data)
         user_id = (User.objects.filter(email=validated_data['email']).values('id'))[0]['id']
@@ -53,10 +53,11 @@ class EditSerializer(serializers.ModelSerializer):
     username = serializers.CharField(validators=[UniqueValidator(queryset=User.objects.all())]) 
     skills = serializers.SerializerMethodField()
     genres = serializers.SerializerMethodField()
+    artists = serializers.SerializerMethodField()
 
     class Meta:
         model=User
-        fields=('username','email','password','skills','genres')
+        fields=('username','email','password','skills','genres','artists')
 
     def get_skills(self, obj):
         id = self.context.get("id")
@@ -67,6 +68,11 @@ class EditSerializer(serializers.ModelSerializer):
         id = self.context.get("id")
         genres = self.context.get("genres")        
         return get_list_field(id, None, "genre", genres)
+    
+    def get_artists(self, obj):
+        id = self.context.get("id")
+        artists = self.context.get("artists")
+        return get_list_field(id, None, "artist", artists)
     
     def update(self, instance, validated_data):
         original_email = validated_data.get('email', instance.email)
@@ -81,19 +87,24 @@ class EditSerializer(serializers.ModelSerializer):
         
         instance.save()
         
-        # update skills/genres in corresponding tables
+        # update other fields in corresponding tables
         id = self.context.get('id')
         for field in List_Fields:
             field_name = field.value
             field_list = self.context.get(field_name)
-            if field_name == 'skills':
-                User_Skills.objects.filter(user_id=id).delete()
-                for obj in field_list:    
-                    User_Skills.objects.create(user_id=id, skill_id=obj)
-            elif field_name == 'genres':
-                User_Genres.objects.filter(user_id=id).delete()
-                for obj in field_list:    
-                    User_Genres.objects.create(user_id=id, genre_id=obj)
+            if field_list:
+                if field_name == 'skills':
+                    User_Skills.objects.filter(user_id=id).delete()
+                    for obj in field_list:    
+                        User_Skills.objects.create(user_id=id, skill_id=obj)
+                elif field_name == 'genres':
+                    User_Genres.objects.filter(user_id=id).delete()
+                    for obj in field_list:    
+                        User_Genres.objects.create(user_id=id, genre_id=obj)
+                elif field_name == 'artists':
+                    User_Artists.objects.filter(user_id=id).delete()
+                    for obj in field_list:    
+                        User_Artists.objects.create(user_id=id, artist=obj)
         
         return instance
